faq restructure & clean up
