--- src/pages/FAQ/index.js
@@ -1,11 +1,0 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
-import "./styles/index.css";
-import FAQ from "./FAQ";
-
-const root = ReactDOM.createRoot(document.getElementById("root"));
-root.render(
-  <React.StrictMode>
-    <FAQ />
-  </React.StrictMode>
-);
